backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.7
google-generativeai==0.3.2
google-auth==2.23.4
google-cloud-speech==2.21.0
SpeechRecognition==3.10.0
pyttsx3==2.90
langdetect==1.0.9
requests==2.31.0
cors==1.0.1
fastapi-cors==0.0.6
python-dotenv==1.0.0
bcrypt==4.1.2
email-validator==2.1.0
aiofiles==23.2.1
gtts==2.4.0
pyaudio==0.2.11
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
pandas==2.1.4
jsonschema==4.20.0
websockets==12.0
httpx==0.25.2